refactor: SuggestionsList atoms => molecules로 이동 및 SuggestionItem 컴포넌트 추가
