+++ src/components/index.ts
@@ -1,3 +1,5 @@
-export { SuggestionsList } from 'components/atoms/SuggestionsList';
+export { SuggestionItem } from 'components/atoms/SuggestionItem';
+
+export { SuggestionsList } from 'components/molecules/SuggestionsList';
 
 export { AutoComplete } from 'components/organisms/AutoComplete';
