+++ src/components/molecules/SuggestionsList/style.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const Container = styled.ul`
+  max-height: 400px;
+  overflow-y: auto;
+  background-color: white;
+  display: flex;
+  flex-direction: column;
+  margin: 5px 10px;
+  padding: 15px;
+  gap: 15px;
+  font-size: 16px;
+  border-radius: 20px;
+`;
